requirements.txt
# additional python packages needed
pyyaml  # read yaml files